requirements.txt
kafka-python
redis
cython
jsonobject
-e git+https://github.com/subham6651/jsonobject.git@master#egg=jsonobject